+++ CHANGELOG.md
@@ -5,6 +5,15 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Added
+
+- the `request_opts` options to `SPARQL.Client.query/3` which will be passed as 
+ the `opts` option value to the `Tesla.request/2` function
+
+
+
 ## 0.2.1 - 2018-09-17
 
 ### Fixed